+++ lib/ZarrCore/src/ZArray.jl
@@ -398,7 +398,7 @@ Creates a new empty zarr array with element type `T` and array dimensions `dims`
 * `fill_value=nothing` value to represent missing values
 * `fill_as_missing=false` set to `true` shall fillvalue s be converted to `missing`s
 * `filters`=filters to be applied
-* `compressor` compressor type and properties, defaulting to the installation default `DEFAULT_COMPRESSOR[]`
+* `compressor` compressor type and properties, defaulting to `default_compressor()`
 * `attrs=Dict()` a dict containing key-value pairs with metadata attributes associated to the array
 * `writeable=true` determines if the array is opened in read-only or write mode
 * `indent_json=false` determines if indents are added to format the json files `.zarray` and `.zattrs`.  This makes them more readable, but increases file size.
@@ -434,7 +434,7 @@ function zcreate(::Type{T},storage::AbstractStore,
   chunks=dims,
   fill_value=nothing,
   fill_as_missing=false,
-  compressor=DEFAULT_COMPRESSOR[],
+  compressor=default_compressor(),
   filters = filterfromtype(T), 
   attrs=Dict(),
   writeable=true,